chore: Refactor DistributionTest.t.sol to skip mainnet forking test
Refactor the `DistributionTest.t.sol` contract to include a modifier `skipMainnetForkingTest` that skips the test execution on the mainnet forks. This modifier checks the `block.chainid` and returns if it is equal to 1 (mainnet). The purpose of this change is to isolate the testing to non-mainnet forks until the issue with mainnet forking is resolved.